packages/gate-agent/pyproject.toml
[project]
name = "gate-agent"
version = "0.1.0"
description = "Festival gate counter — buffered MQTT publisher with computer vision"
requires-python = ">=3.12"
dependencies = [
  "paho-mqtt>=2.1.0",
  "pyyaml>=6.0.2",
  "pydantic>=2.9.0",
  "python-ulid>=3.0.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "opencv-python>=4.10.0",
  "numpy>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gate_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
